+++ client/src/data/constants/VIEWS.ts
@@ -37,7 +37,6 @@ export const VIEWS = {
   },
 
   info: {
-    feedback: "feedback",
     howToPlay: "howToPlay",
   },
 }
